[feat]: Move UsageGoal To Core Domain
