+++ data/usagestats/src/main/java/com/hmh/hamyeonham/usagestats/datasource/local/UsageGoalsLocalDataSource.kt
@@ -2,8 +2,9 @@ package com.hmh.hamyeonham.usagestats.datasource.local
 
 import com.hmh.hamyeonham.core.database.dao.UsageGoalsDao
 import com.hmh.hamyeonham.core.database.dao.UsageTotalGoalDao
+import com.hmh.hamyeonham.core.database.model.UsageGoalsEntity
+import com.hmh.hamyeonham.core.domain.usagegoal.model.UsageGoal
 import com.hmh.hamyeonham.usagestats.mapper.toUsageGoal
-import com.hmh.hamyeonham.usagestats.model.UsageGoal
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
@@ -21,5 +22,31 @@ class UsageGoalsLocalDataSource @Inject constructor(
                 emit(result)
             }
         }
+    }
+
+    suspend fun getUsageGoal(packageName: String): UsageGoal {
+        return usageGoalsDao.getUsageGoal(packageName).toUsageGoal()
+    }
+
+    suspend fun addUsageGoal(usageGoal: UsageGoal) {
+        usageGoalsDao.insertUsageGoal(
+            UsageGoalsEntity(
+                usageGoal.packageName,
+                usageGoal.goalTime
+            )
+        )
+    }
+
+    suspend fun addUsageGoalList(usageGoalList: List<UsageGoal>) {
+        usageGoalsDao.insertUsageGoalList(usageGoalList.map {
+            UsageGoalsEntity(
+                it.packageName,
+                it.goalTime
+            )
+        })
+    }
+
+    suspend fun deleteUsageGoal(packageName: String) {
+        usageGoalsDao.deleteByPackageName(packageName)
     }
 }
