+++ data/usagestats/src/main/java/com/hmh/hamyeonham/usagestats/datasource/remote/UsageGoalsRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.hmh.hamyeonham.usagestats.datasource.remote
 
+import com.hmh.hamyeonham.core.domain.usagegoal.model.UsageGoal
 import com.hmh.hamyeonham.core.network.usagegoal.UsageGoalService
 import com.hmh.hamyeonham.usagestats.mapper.toUsageGoalList
-import com.hmh.hamyeonham.usagestats.model.UsageGoal
 import javax.inject.Inject
 
 class UsageGoalsRemoteDataSource @Inject constructor(
@@ -11,9 +11,5 @@ class UsageGoalsRemoteDataSource @Inject constructor(
 
     suspend fun getUsageGoals(): Result<List<UsageGoal>> {
         return runCatching { usageGoalService.getUsageGoal().data.toUsageGoalList() }
-    }
-
-    fun addUsageGoal(usageGoal: UsageGoal) {
-
     }
 }
